dpe: Modify Sign command for better ML-DSA support
This clarifies that the sign command will take `SIGN_DATA` instead of a
hash so it can accept an external mu parameter.

This also adds a new command to sign a non-prehashed message. This is
mostly for platforms that may not support external mu, but still want to
sign with ML-DSA.